requirements.txt
keras==2.12.0
keras_nightly==2.5.0.dev2021032900
matlab==0.1
matlabengine==9.13.7
mediapipe==0.9.2.1
numpy==1.22.3
opencv_python==4.5.4.58
Pillow==9.5.0
scikit_learn==1.2.2
segment_anything==1.0
tensorflow==2.10.0
torch==1.13.0+cu117
